Refactoring Facebook into a service
